# docs/api/apifox-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "网格社区治理平台 API",
    "description": "网格社区治理平台后端接口文档 v1.2.0 — 完整覆盖所有控制器",
    "version": "1.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "本地开发环境"
    }
  ],
...
  "paths": {
...
    "/events": {
      "post": {
        "tags": [
          "事件"
        ],
        "summary": "创建事件",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "get": {
        "tags": [
          "事件"
        ],
        "summary": "分页查询事件列表",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "areaId",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/events/{id}": {
      "get": {
        "tags": [
          "事件"
        ],
        "summary": "事件详情",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/events/by-external/{externalEventId}": {
      "get": {
        "tags": [
          "事件"
        ],
        "summary": "通过外部事件ID获取详情",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "externalEventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/events/{id}/dispatch": {
      "post": {
        "tags": [
          "事件"
        ],
        "summary": "派单",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/events/{id}/ignore": {
      "post": {
        "tags": [
          "事件"
        ],
        "summary": "标记误报",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/events/false-alarms": {
      "get": {
        "tags": [
          "事件"
        ],
        "summary": "误报记录列表",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/events/batch-delete": {
      "post": {
        "tags": [
          "事件"
        ],
        "summary": "批量删除事件",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/events/{id}/urgency": {
      "put": {
        "tags": [
          "事件"
        ],
        "summary": "更新紧急程度",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/events/auto-escalate": {
      "post": {
        "tags": [
          "事件"
        ],
        "summary": "手动触发自动升级",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/events/{id}/close": {
      "put": {
        "tags": [
          "事件"
        ],
        "summary": "关闭事件",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/events/{id}/reopen": {
      "put": {
        "tags": [
          "事件"
        ],
        "summary": "重新打开事件",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/events/{id}/attachments": {
      "get": {
        "tags": [
          "事件"
        ],
        "summary": "获取事件附件",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/events/{id}/timeline": {
      "get": {
        "tags": [
          "事件"
        ],
        "summary": "获取事件生命周期时间轴",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/events/statistics": {
      "get": {
        "tags": [
          "事件"
        ],
        "summary": "事件统计数据",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/events/{id}/audit": {
      "post": {
        "tags": [
          "事件"
        ],
        "summary": "审核事件",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
    "/events/{id}/rate": {
      "post": {
        "tags": [
          "事件"
        ],
        "summary": "群众评价",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}